refactor(stella-model): split anthropic.rs tests into a submodule
anthropic.rs had grown to 2409 lines — 223 over its file-size ratchet
limit of 2186 (allowlisted 2136 + 50 slack) — reddening the gate on every
PR. Move the ~1436-line `#[cfg(test)] mod tests` block into
src/anthropic/tests.rs (`mod tests;`), dropping anthropic.rs to 974 lines.
Because it now sits under MAX_NEW_LINES (1500), remove its allowlist entry
entirely per the ratchet's shrink-and-lock rule; the new tests.rs is 1433
lines, under the 1500 cap for non-allowlisted files.

Pure refactor, no behavior change — the tests are byte-identical and
`use super::*;` resolves to the same parent module from the child. Verified:
cargo fmt --check clean, clippy -p stella-model -D warnings clean, ratchet
OK (259 files), cargo test -p stella-model 185 passed / 0 failed.
